pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Duellis Padilha",email = "felipe.padilha@brasilnetcp.net.br"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "typer (>=0.15.2,<0.16.0)",
    "netmiko (>=4.5.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "api", from = "src"}]


[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "asyncio"
]